# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type WasteStatus = typeof WasteStatus[keyof typeof WasteStatus];


export const WasteStatus = {
  draft: 'draft',
  registered: 'registered',
  collected: 'collected',
  in_storage: 'in_storage',
  assigned_transport: 'assigned_transport',
  in_transit: 'in_transit',
  delivered: 'delivered',
  in_treatment: 'in_treatment',
  disposed: 'disposed',
  completed: 'completed',
} as const;

export type WasteRecordRisk = typeof WasteRecordRisk[keyof typeof WasteRecordRisk];


export const WasteRecordRisk = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface WasteRecord {
  id: string;
  trackingId: string;
  category: string;
  source: string;
  quantity: number;
  unit: string;
  status: WasteStatus;
  createdAt: string;
  dueDate: string;
  /** @nullable */
  handler?: string | null;
  /** @nullable */
  facility?: string | null;
  risk?: WasteRecordRisk;
}

export type WasteRecordInputRisk = typeof WasteRecordInputRisk[keyof typeof WasteRecordInputRisk];


export const WasteRecordInputRisk = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface WasteRecordInput {
  category: string;
  source: string;
  /** @minimum 0 */
  quantity: number;
  unit: string;
  dueDate: string;
  handler?: string;
  facility?: string;
  risk?: WasteRecordInputRisk;
}

export type WasteRecordUpdateRisk = typeof WasteRecordUpdateRisk[keyof typeof WasteRecordUpdateRisk];


export const WasteRecordUpdateRisk = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface WasteRecordUpdate {
  category?: string;
  source?: string;
  /** @minimum 0 */
  quantity?: number;
  unit?: string;
  dueDate?: string;
  handler?: string;
  facility?: string;
  risk?: WasteRecordUpdateRisk;
}

export interface StatusChangeInput {
  status: WasteStatus;
  note?: string;
}

export interface TimelineEntry {
  id: string;
  status: WasteStatus;
  label: string;
  actor: string;
  timestamp: string;
  /** @nullable */
  note?: string | null;
}

export interface Document {
  id: string;
  name: string;
  type: string;
  uploadedAt: string;
  /** @nullable */
  size?: string | null;
}

export type WasteRecordDetail = WasteRecord & {
  timeline: TimelineEntry[];
  documents: Document[];
};

export interface WasteRecordList {
  items: WasteRecord[];
  total: number;
  page: number;
  pageSize: number;
}

export type ActivityTone = typeof ActivityTone[keyof typeof ActivityTone];


export const ActivityTone = {
  neutral: 'neutral',
  success: 'success',
  warning: 'warning',
  danger: 'danger',
  info: 'info',
} as const;

export interface Activity {
  id: string;
  actor: string;
  action: string;
  target: string;
  timestamp: string;
  tone: ActivityTone;
}

export type DashboardMetrics = {
  totalGenerated: number;
  pendingCollection: number;
  inTransit: number;
  pendingDisposal: number;
  completed: number;
  overdue: number;
};

export type DashboardStatusBreakdownItem = {
  status: WasteStatus;
  count: number;
};

export type DashboardCompliance = {
  score: number;
  missingDocuments: number;
  expiringSoon: number;
};

export interface Dashboard {
  metrics: DashboardMetrics;
  statusBreakdown: DashboardStatusBreakdownItem[];
  activity: Activity[];
  compliance: DashboardCompliance;
}

export type SearchParameter = string;

export type StatusParameter = WasteStatus;

export type PageParameter = number;

export type PageSizeParameter = number;

export type ListWasteRecordsParams = {
search?: SearchParameter;
status?: StatusParameter;
/**
 * @minimum 1
 */
page?: PageParameter;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: PageSizeParameter;
};

export type ListActivityParams = {
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: PageSizeParameter;
};