# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Operations overview
  - name: waste
    description: Hazardous waste lifecycle
  - name: activity
    description: Audit and lifecycle activity
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard overview
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /waste-records:
    get:
      operationId: listWasteRecords
      tags: [waste]
      summary: List waste records
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Waste records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteRecordList"
    post:
      operationId: createWasteRecord
      tags: [waste]
      summary: Register a waste record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WasteRecordInput"
      responses:
        "201":
          description: Waste record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteRecord"
  /waste-records/{id}:
    get:
      operationId: getWasteRecord
      tags: [waste]
      summary: Get a waste record
      parameters:
        - $ref: "#/components/parameters/RecordId"
      responses:
        "200":
          description: Waste record details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteRecordDetail"
        "404":
          description: Record not found
    patch:
      operationId: updateWasteRecord
      tags: [waste]
      summary: Update a waste record
      parameters:
        - $ref: "#/components/parameters/RecordId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WasteRecordUpdate"
      responses:
        "200":
          description: Waste record updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteRecord"
  /waste-records/{id}/status:
    post:
      operationId: changeWasteStatus
      tags: [waste]
      summary: Advance a waste record through its lifecycle
      parameters:
        - $ref: "#/components/parameters/RecordId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusChangeInput"
      responses:
        "200":
          description: Updated waste record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteRecordDetail"
  /activity:
    get:
      operationId: listActivity
      tags: [activity]
      summary: List recent operational activity
      parameters:
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
components:
  parameters:
    RecordId:
      name: id
      in: path
      required: true
      schema:
        type: string
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
    Status:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/WasteStatus"
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    WasteStatus:
      type: string
      enum: [draft, registered, collected, in_storage, assigned_transport, in_transit, delivered, in_treatment, disposed, completed]
    WasteRecord:
      type: object
      required: [id, trackingId, category, source, quantity, unit, status, createdAt, dueDate]
      properties:
        id:
          type: string
        trackingId:
          type: string
        category:
          type: string
        source:
          type: string
        quantity:
          type: number
        unit:
          type: string
        status:
          $ref: "#/components/schemas/WasteStatus"
        createdAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date
        handler:
          type: ["string", "null"]
        facility:
          type: ["string", "null"]
        risk:
          type: string
          enum: [low, medium, high, critical]
    WasteRecordInput:
      type: object
      required: [category, source, quantity, unit, dueDate]
      properties:
        category:
          type: string
        source:
          type: string
        quantity:
          type: number
          minimum: 0
        unit:
          type: string
        dueDate:
          type: string
          format: date
        handler:
          type: string
        facility:
          type: string
        risk:
          type: string
          enum: [low, medium, high, critical]
    WasteRecordUpdate:
      type: object
      properties:
        category:
          type: string
        source:
          type: string
        quantity:
          type: number
          minimum: 0
        unit:
          type: string
        dueDate:
          type: string
          format: date
        handler:
          type: string
        facility:
          type: string
        risk:
          type: string
          enum: [low, medium, high, critical]
    StatusChangeInput:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/WasteStatus"
        note:
          type: string
    WasteRecordDetail:
      allOf:
        - $ref: "#/components/schemas/WasteRecord"
        - type: object
          required: [timeline, documents]
          properties:
            timeline:
              type: array
              items:
                $ref: "#/components/schemas/TimelineEntry"
            documents:
              type: array
              items:
                $ref: "#/components/schemas/Document"
    WasteRecordList:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WasteRecord"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    TimelineEntry:
      type: object
      required: [id, status, label, actor, timestamp]
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/WasteStatus"
        label:
          type: string
        actor:
          type: string
        timestamp:
          type: string
          format: date-time
        note:
          type: ["string", "null"]
    Document:
      type: object
      required: [id, name, type, uploadedAt]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        uploadedAt:
          type: string
          format: date-time
        size:
          type: ["string", "null"]
    Activity:
      type: object
      required: [id, actor, action, target, timestamp, tone]
      properties:
        id:
          type: string
        actor:
          type: string
        action:
          type: string
        target:
          type: string
        timestamp:
          type: string
          format: date-time
        tone:
          type: string
          enum: [neutral, success, warning, danger, info]
    Dashboard:
      type: object
      required: [metrics, statusBreakdown, activity, compliance]
      properties:
        metrics:
          type: object
          required: [totalGenerated, pendingCollection, inTransit, pendingDisposal, completed, overdue]
          properties:
            totalGenerated:
              type: number
            pendingCollection:
              type: integer
            inTransit:
              type: integer
            pendingDisposal:
              type: integer
            completed:
              type: integer
            overdue:
              type: integer
        statusBreakdown:
          type: array
          items:
            type: object
            required: [status, count]
            properties:
              status:
                $ref: "#/components/schemas/WasteStatus"
              count:
                type: integer
        activity:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        compliance:
          type: object
          required: [score, missingDocuments, expiringSoon]
          properties:
            score:
              type: number
            missingDocuments:
              type: integer
            expiringSoon:
              type: integer