# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityTone } from './activityTone';

export interface Activity {
  id: string;
  actor: string;
  action: string;
  target: string;
  timestamp: Date;
  tone: ActivityTone;
}