# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { DashboardCompliance } from './dashboardCompliance';
import type { DashboardMetrics } from './dashboardMetrics';
import type { DashboardStatusBreakdownItem } from './dashboardStatusBreakdownItem';

export interface Dashboard {
  metrics: DashboardMetrics;
  statusBreakdown: DashboardStatusBreakdownItem[];
  activity: Activity[];
  compliance: DashboardCompliance;
}