# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: string;
  name: string;
  type: string;
  uploadedAt: Date;
  /** @nullable */
  size?: string | null;
}