# lib/api-zod/src/generated/types/statusChangeInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WasteStatus } from './wasteStatus';

export interface StatusChangeInput {
  status: WasteStatus;
  note?: string;
}