# lib/api-zod/src/generated/types/timelineEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WasteStatus } from './wasteStatus';

export interface TimelineEntry {
  id: string;
  status: WasteStatus;
  label: string;
  actor: string;
  timestamp: Date;
  /** @nullable */
  note?: string | null;
}