# lib/api-zod/src/generated/types/wasteRecord.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WasteRecordRisk } from './wasteRecordRisk';
import type { WasteStatus } from './wasteStatus';

export interface WasteRecord {
  id: string;
  trackingId: string;
  category: string;
  source: string;
  quantity: number;
  unit: string;
  status: WasteStatus;
  createdAt: Date;
  dueDate: Date;
  /** @nullable */
  handler?: string | null;
  /** @nullable */
  facility?: string | null;
  risk?: WasteRecordRisk;
}