# lib/api-zod/src/generated/types/wasteRecordDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Document } from './document';
import type { TimelineEntry } from './timelineEntry';
import type { WasteRecord } from './wasteRecord';

export type WasteRecordDetail = WasteRecord & {
  timeline: TimelineEntry[];
  documents: Document[];
};