# lib/api-zod/src/generated/types/wasteRecordList.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WasteRecord } from './wasteRecord';

export interface WasteRecordList {
  items: WasteRecord[];
  total: number;
  page: number;
  pageSize: number;
}