# lib/api-zod/src/generated/types/wasteRecordUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WasteRecordUpdateRisk } from './wasteRecordUpdateRisk';

export interface WasteRecordUpdate {
  category?: string;
  source?: string;
  /** @minimum 0 */
  quantity?: number;
  unit?: string;
  dueDate?: Date;
  handler?: string;
  facility?: string;
  risk?: WasteRecordUpdateRisk;
}