# lib/api-zod/src/generated/types/wasteStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type WasteStatus = typeof WasteStatus[keyof typeof WasteStatus];


export const WasteStatus = {
  draft: 'draft',
  registered: 'registered',
  collected: 'collected',
  in_storage: 'in_storage',
  assigned_transport: 'assigned_transport',
  in_transit: 'in_transit',
  delivered: 'delivered',
  in_treatment: 'in_treatment',
  disposed: 'disposed',
  completed: 'completed',
} as const;